ai-service/requirements.txt
flask==3.0.0
python-dotenv==1.0.1
requests==2.31.0
moviepy==1.0.3
flask-cors==4.0.0
